# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: IFRN Editorial Portal API — Fase 1
  description: >-
    API do ifrn-editorial-portal, escopo da Fase 1 (caminho crítico de
    leitura): health check, autenticação, sessão, autorização e leitura de
    um documento Markdown de demonstração no repositório
    cte-zl-ifrn/central-ajuda. Não inclui edição, upload, criação de
    branches/commits/Pull Requests, conforme docs/phase-1-plan.md.
  version: "0.1.0"
servers:
  - url: http://localhost:8000
    description: Backend local (desenvolvimento)
...
paths:
...
  /auth/callback:
    get:
      tags: [auth]
      summary: Callback do OAuth do GitHub
      description: >-
        Valida `state` e `code`, identifica o usuário, verifica sua
        permissão no repositório `cte-zl-ifrn/central-ajuda` e cria a
        sessão do portal.
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: Código de autorização emitido pelo GitHub
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: Valor de state gerado em /auth/login
      responses:
        "302":
          description: >-
            Sessão criada; redirecionamento de volta para o frontend, com
            cookie de sessão definido.
          headers:
            Set-Cookie:
              schema:
                type: string
            Location:
              schema:
                type: string
        "400":
          description: state ausente, inválido ou expirado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Falha de comunicação com o GitHub durante o callback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: portal_session
      description: >-
        Cookie de sessão HttpOnly/Secure criado após /auth/callback. O
        nome exato do cookie é definido pela implementação do backend
        (ver backend/README.md).
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Código de erro estável, sem detalhes internos
          example: "unauthenticated"
        message:
          type: string
          description: Mensagem legível para o usuário final
        correlation_id:
          type: string
          description: Identificador de correlação para rastreamento em logs